# swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "all api documation commonApp nodeJs",
        "version": "1.0.0",
        "title": "Swagger of UMS",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "mansuri.11300406@lpu.in"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "host": "safe-atoll-64757.herokuapp.com",
    "basePath": "/v1",
    "produces": [
        "application/json"
    ],
    "paths": {
        "/users/login": {
            "post": {
                "summary": "user login",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Common"
                ],
                "parameters": [{
                    "in": "body",
                    "name": "user",
                    "description": "The user login.",
                    "schema": {
                        "type": "object",
                        "required": [
                            "register_id",
                            "password"
                        ],
                        "properties": {
                            "register_id": {
                                "type": "string"
                            },
                            "password": {
                                "type": "string"
                            }
                        }
                    }
                }],
                "responses": {}
            }
        },
        "/users/create-student": {
            "post": {
                "summary": "create_student",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Admins"
                ],
                "parameters": [{
                        "name": "Authorization",
                        "in": "header",
                        "description": "an authorization header, Please add Bearer keyword before token",
                        "required": "true",
                        "type": "string"
                    },  
                    {
                        "in": "body",
                        "name": "user",
                        "description": "The user register.",
                        "schema": {
                            "type": "object",
                            "required": [
                                "register_id",
                                "password"
                            ],
                            "properties": {
                                "register_id": {
                                    "type": "string"
                                },
                                "password": {
                                    "type": "string"
                                },
                                "first_name": {
                                    "type": "string"
                                },
                                "last_name": {
                                    "type": "string"
                                },
                                "program_id": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/users/create-faculty": {
            "post": {
                "summary": "create faculty",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Admins"
                ],
                "parameters": [{
                        "name": "Authorization",
                        "in": "header",
                        "description": "an authorization header, Please add Bearer keyword before token",
                        "required": "true",
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "user",
                        "description": "The faculty register.",
                        "schema": {
                            "type": "object",
                            "required": [
                                "register_id",
                                "password"
                            ],
                            "properties": {
                                "register_id": {
                                    "type": "string"
                                },
                                "password": {
                                    "type": "string"
                                },
                                "first_name": {
                                    "type": "string"
                                },
                                "last_name": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/users/create-program": {
            "post": {
                "summary": "create program",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Admins"
                ],
                "parameters": [{
                        "name": "Authorization",
                        "in": "header",
                        "description": "an authorization header, Please add Bearer keyword before token",
                        "required": "true",
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "user",
                        "description": "create programs",
                        "schema": {
                            "type": "object",
                            "required": [
                                "register_id",
                                "password"
                            ],
                            "properties": {
                                "name": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/users/get-all-students": {
            "get": {
                "x-swagger-router-controller": "bar",
                "description": "",
                "tags": [
                    "Admins"
                ],
                "parameters": [{
                        "name": "Authorization",
                        "in": "header",
                        "description": "an authorization header, Please add Bearer keyword before token",
                        "required": "true",
                        "type": "string"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "page",
                        "required": false,
                        "type": "number"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "limit",
                        "required": false,
                        "type": "number"
                    },
                    {
                        "name": "sortBy",
                        "in": "query",
                        "description": "sort by",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "q",
                        "in": "query",
                        "description": "for search",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {}
            }
        },
        "/users/get-all-faculties": {
            "get": {
                "x-swagger-router-controller": "bar",
                "description": "",
                "tags": [
                    "Admins"
                ],
                "parameters": [{
                        "name": "Authorization",
                        "in": "header",
                        "description": "an authorization header, Please add Bearer keyword before token",
                        "required": "true",
                        "type": "string"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "page",
                        "required": false,
                        "type": "number"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "limit",
                        "required": false,
                        "type": "number"
                    },
                    {
                        "name": "sortBy",
                        "in": "query",
                        "description": "sort by",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "q",
                        "in": "query",
                        "description": "for search",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {}
            }
        },
        "/users/logout": {
            "get": {
                "summary": "user logout",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Common"
                ],
                "parameters": [{
                    "name": "Authorization",
                    "in": "header",
                    "description": "an authorization header, Please add Bearer keyword before token",
                    "required": "true",
                    "type": "string"
                }],
                "responses": {}
            }
        },
        "/users/get-time-table": {
            "get": {
                "summary": "get time table",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Common"
                ],
                "parameters": [{
                    "name": "Authorization",
                    "in": "header",
                    "description": "an authorization header, Please add Bearer keyword before token",
                    "required": "true",
                    "type": "string"
                }],
                "responses": {}
            }
        },
        "/users/get-exam-time-table": {
            "get": {
                "summary": "get time table",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Common"
                ],
                "parameters": [{
                    "name": "Authorization",
                    "in": "header",
                    "description": "an authorization header, Please add Bearer keyword before token",
                    "required": "true",
                    "type": "string"
                }],
                "responses": {}
            }
        },
...
        "/users/change-password/": {
            "put": {
                "summary": "user change password",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Common"
                ],
                "parameters": [{
                        "name": "Authorization",
                        "in": "header",
                        "description": "an authorization header, Please add Bearer keyword before token",
                        "required": "true",
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "user",
                        "description": "The user login",
                        "schema": {
                            "type": "object",
                            "required": [
                                "old_password",
                                "new_password",
                                "confirm_password"
                            ],
                            "properties": {
                                "old_password": {
                                    "type": "string"
                                },
                                "new_password": {
                                    "type": "string"
                                },
                                "confirm_password": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/users/forgot-password": {
            "post": {
                "summary": "user forgot password",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Common"
                ],
                "parameters": [{
                    "in": "body",
                    "name": "user",
                    "description": "The user login",
                    "schema": {
                        "type": "object",
                        "required": [
                            "email",
                            "password"
                        ],
                        "properties": {
                            "email": {
                                "type": "string"
                            }
                        }
                    }
                }],
                "responses": {}
            }
        },
        "/users/": {
            "get": {
                "summary": "get user deatils",
                "x-swagger-router-controller": "bar",
                "operationId": "impossible",
                "tags": [
                    "Common"
                ],
                "description": "",
                "parameters": [{
                    "name": "Authorization",
                    "in": "header",
                    "description": "an authorization header, Please add Bearer keyword before token",
                    "required": "true",
                    "type": "string"
                }],
                "responses": {}
            },
            "put": {
                "summary": "update user",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Common"
                ],
                "description": "set user skills:- {skill_level:1}.. 'BEGGINER': 1,'INTERMEDIATE': 2,'ADVANCED': 3   \n set theme:- {theme_type: 1} 'WHITE:1, BLACK:2",
                "parameters": [{
                        "name": "Authorization",
                        "in": "header",
                        "description": "an authorization header, Please add Bearer keyword before token",
                        "required": "true",
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "user",
                        "description": "The user update",
                        "schema": {
                            "type": "object",
                            "required": [],
                            "properties": {
                                "first_name": {
                                    "type": "string"
                                },
                                "last_name": {
                                    "type": "string"
                                },
                                "father_name": {
                                    "type": "string"
                                },
                                "mother_name": {
                                    "type": "string"
                                },
                                "mobile": {
                                    "type": "string"
                                },
                                "date_of_birth": {
                                    "type": "number"
                                },
                                "gender": {
                                    "type": "string"
                                },
                                "email": {
                                    "type": "string"
                                },
                                "address": {
                                    "type": "string"
                                },
                                "country_id": {
                                    "type": "string"
                                },
                                "state_id": {
                                    "type": "string"
                                },
                                "ciry_id": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {}
            }
        },
...
        "/users/contact-us": {
            "post": {
                "summary": "user can add query",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "parameters": [{
                        "name": "Authorization",
                        "in": "header",
                        "description": "an authorization header, Please add Bearer keyword before token",
                        "required": "true",
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "user",
                        "description": "add contactUs query",
                        "schema": {
                            "type": "object",
                            "required": [
                                "email",
                                "username",
                                "query"
                            ],
                            "properties": {
                                "email": {
                                    "type": "string"
                                },
                                "username": {
                                    "type": "string"
                                },
                                "query": {
                                    "type": "string"
                                },
                                "type": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {}
            }
        },
...
        "/users/create-course": {
            "post": {
                "summary": "create course",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Admins"
                ],
                "parameters": [{
                        "name": "Authorization",
                        "in": "header",
                        "description": "an authorization header, Please add Bearer keyword before token",
                        "required": "true",
                        "type": "string"
                    },

                    {
                        "in": "body",
                        "name": "user",
                        "description": "The create course.",
                        "schema": {
                            "type": "object",
                            "required": [],
                            "properties": {
                                "code": {
                                    "type": "string"
                                },
                                "name": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {}
            }
        },
...
        "/users/student-course-list": {
            "get": {
                "summary": "get user's selected courses",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "parameters": [{
                    "name": "Authorization",
                    "in": "header",
                    "description": "an authorization header, Please add Bearer keyword before token",
                    "required": "true",
                    "type": "string"
                }],
                "responses": {}
            }
        },
        "/users/faculty-course-list": {
            "get": {
                "summary": "get faculty's assigned courses",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Faculty"
                ],
                "parameters": [{
                    "name": "Authorization",
                    "in": "header",
                    "description": "an authorization header, Please add Bearer keyword before token",
                    "required": "true",
                    "type": "string"
                }],
                "responses": {}
            }
        },
...
    }
}